Move typeahead to separate module
